+++ src/main/java/shop/bookbom/shop/domain/cart/dto/repsonse/CartResponse.java
@@ -1,0 +1,11 @@
+package shop.bookbom.shop.domain.cart.dto.repsonse;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class CartResponse {
+    private Long bookId;
+    private int quantity;
+}
