+++ src/main/java/shop/bookbom/shop/domain/cart/repository/CartRepository.java
@@ -8,6 +8,6 @@ import shop.bookbom.shop.domain.cart.entity.Cart;
 import shop.bookbom.shop.domain.member.entity.Member;
 
 public interface CartRepository extends JpaRepository<Cart, Long> {
-    @Query("select c from Cart c join fetch CartItem ci where c.member = :member")
-    Optional<Cart> getCartByMemberFetch(Member member);
+    @Query("select c from Cart c join fetch c.cartItems ci where c.member = :member")
+    Optional<Cart> getCartFetchItems(Member member);
 }
