+++ src/main/java/shop/bookbom/shop/domain/cart/service/CartFindService.java
@@ -16,7 +16,7 @@ public class CartFindService {
 
     @Transactional
     public Cart getCart(Member member) {
-        return cartRepository.getCartByMemberFetch(member)
+        return cartRepository.getCartFetchItems(member)
                 .orElseGet(() -> cartRepository.save(Cart.builder()
                         .member(member)
                         .build()));
